backend/requirements.txt
flask
torch
numpy
opencv-python
scikit-image
pillow